Fix up defect where widget doesn't pick up a finished or reset state correctly, if the state was changed from the app
Widget would be ok if it was ticking, but not for the other states. Only reproducible on a device.
Seems as though the device might keep the widget in memory for a short time, and not unload it, hence viewDidLoad doesn't get called every time. Moving refreshUI to viewWillAppear solves the problem.